Working code with intermediate changes. Committing for safe keeping. Now going to add the logic for migrating to the smallest memory footprint ring 0 machine.

+++ datacenter.h
@@ -12,7 +12,7 @@
 using namespace std;
 
 //typedef unsigned long EnergyUnit; // In Watt-seconds
-typedef float EnergyUnit; // In Watt-seconds
+typedef double EnergyUnit; // In Watt-seconds
 
 class DataCenter
 {
